[REFACTOR]: отдельные файлы для класса игрока и врага
